+++ pages/signup.tsx
@@ -1,0 +1,103 @@
+import React from "react";
+
+type Props = {};
+
+const SignupPage = (props: Props) => {
+    return (
+        <div className="min-h-full flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
+            <div className="max-w-md w-full space-y-8">
+                <div>
+                    <img
+                        className="mx-auto h-12 w-auto"
+                        src="https://tailwindui.com/img/logos/workflow-mark-indigo-600.svg"
+                        alt="Workflow"
+                    />
+                    <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
+                        Sign up !
+                    </h2>
+                </div>
+                <form className="mt-8 space-y-6" action="#" method="POST">
+                    <input type="hidden" name="remember" defaultValue="true" />
+                    <div className="rounded-md shadow-sm -space-y-px">
+                        <div>
+                            <label
+                                htmlFor="email-address"
+                                className="text-base text-gray-500"
+                            >
+                                Username
+                            </label>
+                            <input
+                                id="username"
+                                name="username"
+                                type="text"
+                                autoComplete="username"
+                                required
+                                className="appearance-none relative block w-full px-3 py-2 border border-gray-400 placeholder-gray-500 text-gray-900 rounded focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm my-2"
+                            />
+                        </div>
+                        <div>
+                            <label
+                                htmlFor="email-address"
+                                className="text-base text-gray-500"
+                            >
+                                Email address
+                            </label>
+                            <input
+                                id="email-address"
+                                name="email"
+                                type="email"
+                                autoComplete="email"
+                                required
+                                className="appearance-none relative block w-full px-3 py-2 border border-gray-400 placeholder-gray-500 text-gray-900 rounded focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm my-2"
+                            />
+                        </div>
+                        <div>
+                            <label htmlFor="password" className="text-base text-gray-500">
+                                Password
+                            </label>
+                            <input
+                                id="password"
+                                name="password"
+                                type="password"
+                                autoComplete="current-password"
+                                required
+                                className="appearance-none relative block w-full px-3 py-2 border border-gray-400 placeholder-gray-500 text-gray-900 rounded focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm my-2"
+                            />
+                        </div>
+                    </div>
+                    <div className="flex items-center justify-between">
+                        <div className="flex items-center">
+                            <input
+                                id="remember-me"
+                                name="remember-me"
+                                type="checkbox"
+                                className="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 rounded"
+                            />
+                            <label
+                                htmlFor="remember-me"
+                                className="ml-2 block text-sm text-gray-900"
+                            >
+                                Remember me
+                            </label>
+                        </div>
+                        <div className="text-sm">
+                            <a href="#" className="font-medium text-gray-900">
+                                Forgot your password?
+                            </a>
+                        </div>
+                    </div>
+                    <div>
+                        <button
+                            type="submit"
+                            className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-black hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+                        >
+                            Sign up
+                        </button>
+                    </div>
+                </form>
+            </div>
+        </div>
+    );
+};
+
+export default SignupPage;
